+++ src/ModuleManager/ModuleListViewController.m
@@ -388,7 +388,7 @@
 			int len = [selectedRows count];
             ModuleListObject *mlo = nil;
 			if(len > 0) {
-				unsigned int indexes[len];
+				NSUInteger indexes[len];
 				[selectedRows getIndexes:indexes maxCount:len inIndexRange:nil];
 				
 				for(int i = 0;i < len;i++) {
@@ -406,7 +406,7 @@
 	}
 }
 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item {
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item {
     int count = 0;
 	
 	if(item == nil) {
@@ -417,7 +417,7 @@
 	return count;
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item {
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item {
     
     // we only have modules here
     NSDictionary *ret = nil;
