+++ src/UserInterface/GenBookViewController.m
@@ -373,7 +373,7 @@
 			NSMutableArray *sel = [NSMutableArray arrayWithCapacity:len];
             SwordModuleTreeEntry *item = nil;
 			if(len > 0) {
-				unsigned int indexes[len];
+				NSUInteger indexes[len];
 				[selectedRows getIndexes:indexes maxCount:len inIndexRange:nil];
 				
 				for(int i = 0;i < len;i++) {
@@ -402,7 +402,7 @@
 }
 
 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item {
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item {
     int count = 0;
 	
 	if(item == nil) {
@@ -416,7 +416,7 @@
 	return count;
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item {
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item {
     
     SwordModuleTreeEntry *ret = nil;
     if(item == nil) {
